+++ src/GSvar/ProcessedSampleWidget.cpp
@@ -177,6 +177,8 @@ void ProcessedSampleWidget::updateGUI()
 	ui_->run->setText("<a href=\"" + run + "\">"+run+"</a>");
 	ui_->merged->setText(mergedSamples());
 	ui_->lab_operator->setText(ps_data.lab_operator);
+	ui_->processing_modus->setText(ps_data.processing_modus);
+	ui_->batch_number->setText(ps_data.batch_number);
 	ui_->processing_input->setText(ps_data.processing_input);
 	ui_->molarity->setText(ps_data.molarity);
 	QString normal_sample = ps_data.normal_sample_name;